Satisfies problem 02
Bank keeps deposits in a synchronized collection and knows how to
sum the balance of that collection. Teller knows how to request
a balance and deposit an amount. Bios uses formal cli to choose
which role to adopt and what a teller may ask to do. Pom adds
apache cli dependency.